feat : 경로 조회 시 노선 정보 추가로 반환

+++ src/main/java/nextstep/subway/applicaion/dto/LineRequest.java
@@ -9,7 +9,7 @@ import nextstep.subway.domain.Line;
 public class LineRequest {
     private String name;
     private String color;
-    private long extraFare;
+    private int extraFare;
     private Long upStationId;
     private Long downStationId;
     private int distance;
